+++ installation_and_upgrade/ibex_install_utils/install_tasks.py
@@ -2,120 +2,19 @@
 Tasks associated with install
 """
 from __future__ import division, unicode_literals, print_function
-import pprint
-import tempfile
-import zipfile
-from contextlib import contextmanager, closing
 
 import os
-import shutil
-import socket
-import subprocess
-from datetime import date, datetime
-from time import sleep
 
-import psutil
-import git
-
-from ibex_install_utils.admin_runner import AdminCommandBuilder
-from ibex_install_utils.ca_utils import CaWrapper
-from ibex_install_utils.exceptions import UserStop, ErrorInRun, ErrorInTask
 from ibex_install_utils.file_utils import FileUtils
-from ibex_install_utils.run_process import RunProcess
-from ibex_install_utils.task import task
-from ibex_install_utils.user_prompt import UserPrompt
-from ibex_install_utils.motor_params import get_params_and_save_to_file
-from ibex_install_utils.kafka_utils import add_required_topics
-
-try:
-    from subprocess import DETACHED_PROCESS
-except ImportError:
-    # For Py2 compatibility, can be removed once we are on Py3.
-    DETACHED_PROCESS = 0x00000008
-
-BACKUP_DATA_DIR = os.path.join("C:\\", "data")
-BACKUP_DIR = os.path.join(BACKUP_DATA_DIR, "old")
-
-INSTRUMENT_BASE_DIR = os.path.join("C:\\", "Instrument")
-APPS_BASE_DIR = os.path.join(INSTRUMENT_BASE_DIR, "Apps")
-EPICS_PATH = os.path.join(APPS_BASE_DIR, "EPICS")
-
-SYSTEM_SETUP_PATH = os.path.join(EPICS_PATH, "SystemSetup")
-GUI_PATH = os.path.join(APPS_BASE_DIR, "Client")
-GUI_PATH_E4 = os.path.join(APPS_BASE_DIR, "Client_E4")
-PYTHON_PATH = os.path.join(APPS_BASE_DIR, "Python")
-PYTHON_3_PATH = os.path.join(APPS_BASE_DIR, "Python3")
-CONFIG_UPGRADE_SCRIPT_DIR = os.path.join(EPICS_PATH, "misc", "upgrade", "master")
-EPICS_UTILS_PATH = os.path.join(APPS_BASE_DIR, "EPICS_UTILS")
-DESKTOP_TRAINING_FOLDER_PATH = os.path.join(os.environ["userprofile"], "desktop", "Mantid+IBEX training")
-SETTINGS_CONFIG_FOLDER = os.path.join("Settings", "config")
-SETTINGS_CONFIG_PATH = os.path.join(INSTRUMENT_BASE_DIR, SETTINGS_CONFIG_FOLDER)
-CALIBRATION_PATH = os.path.join(SETTINGS_CONFIG_PATH, "common")
-SOURCE_FOLDER = os.path.join(os.path.dirname(os.path.realpath(__file__)), "resources")
-SOURCE_MACHINE_SETTINGS_CONFIG_PATH = os.path.join(SOURCE_FOLDER, SETTINGS_CONFIG_FOLDER, "NDXOTHER")
-SOURCE_MACHINE_SETTINGS_COMMON_PATH = os.path.join(SOURCE_FOLDER, SETTINGS_CONFIG_FOLDER, "common")
-
-INST_SHARE_AREA = os.path.join(r"\\isis.cclrc.ac.uk", "inst$")
-
-MYSQL8_INSTALL_DIR = os.path.join(APPS_BASE_DIR, "MySQL")
-MYSQL57_INSTALL_DIR = os.path.join("C:\\", "Program Files", "MySQL", "MySQL Server 5.7")
-MYSQL_LATEST_VERSION = "8.0.19"
-MYSQL_ZIP = os.path.join(INST_SHARE_AREA, "kits$", "CompGroup", "ICP", "MySQL",
-                         "mysql-{}-winx64.zip".format(MYSQL_LATEST_VERSION))
-
-VAR_DIR = os.path.join(INSTRUMENT_BASE_DIR, "var")
-MYSQL_FILES_DIR = os.path.join(VAR_DIR, "mysql")
-PV_BACKUPS_DIR = os.path.join(VAR_DIR, "deployment_pv_backups")
-SQLDUMP_FILE_TEMPLATE = "ibex_db_sqldump_{}.sql"
-
-VCRUNTIME140 = os.path.join("C:\\", "Windows", "System32", "vcruntime140_1.dll")
-VCRUNTIME140_INSTALLER = os.path.join(INST_SHARE_AREA, "kits$", "CompGroup", "ICP", "vcruntime140_installer",
-                                      "vc_redist.x64.exe")
-
-REMOTE_VHD_SRC_DIR = os.path.join(INST_SHARE_AREA, "Kits$", "CompGroup", "chris")
-REMOTE_VHD_DEST_DIR = os.path.join(INST_SHARE_AREA, "Kits$", "CompGroup", "ICP", "VHDS")
-LOCAL_VHD_DIR = os.path.join("C:\\", "Instrument", "VHDS")
-
-
-FILE_TO_REQUEST_VHD_MOUNTING = os.path.join(INSTRUMENT_BASE_DIR, "ibex_vhd_deployment_mount_vhds.txt")
-FILE_TO_REQUEST_VHD_DISMOUNTING = os.path.join(INSTRUMENT_BASE_DIR, "ibex_vhd_deployment_dismount_vhds.txt")
-VHD_MOUNT_DISMOUNT_TIMEOUT = 300
-
-
-class Vhd(object):
-    def __init__(self, name, source_filename, dest_filename, mount_point):
-        self.name = name
-        self.source_filename = source_filename
-        self.dest_filename = dest_filename
-        self.mount_point = mount_point
-
-
-VHDS = [
-    # Key = VHD location, Value = Mount point
-    Vhd(name="Apps", source_filename="empty_apps.vhdx", dest_filename="apps.vhdx", mount_point=APPS_BASE_DIR),
-    Vhd(name="Settings", source_filename="empty_settings.vhdx", dest_filename="settings.vhdx", mount_point=SETTINGS_CONFIG_PATH),
-    Vhd(name="Var", source_filename="empty_var.vhdx", dest_filename="var.vhdx", mount_point=VAR_DIR),
-]
+from ibex_install_utils.tasks.backup_tasks import BackupTasks
+from ibex_install_utils.tasks.client_tasks import ClientTasks
+from ibex_install_utils.tasks.mysql_tasks import MysqlTasks
+from ibex_install_utils.tasks.python_tasks import PythonTasks
+from ibex_install_utils.tasks.server_tasks import ServerTasks
+from ibex_install_utils.tasks.system_tasks import SystemTasks
+from ibex_install_utils.tasks.vhd_tasks import VHDTasks
 
 LABVIEW_DAE_DIR = os.path.join("C:\\", "LabVIEW modules", "DAE")
-
-USER_START_MENU = os.path.join("C:\\", "users", "spudulike", "AppData", "Roaming", "Microsoft", "Windows", "Start Menu")
-PC_START_MENU = os.path.join("C:\\", "ProgramData", "Microsoft", "Windows", "Start Menu")
-SECI = "SECI User interface.lnk"
-SECI_ONE_PATH = os.path.join("C:\\", "Program Files (x86)", "CCLRC ISIS Facility")
-AUTOSTART_LOCATIONS = [os.path.join(USER_START_MENU, "Programs", "Startup", SECI),
-                       os.path.join(PC_START_MENU, "Programs", "Startup", SECI)]
-
-STAGE_DELETED = os.path.join(INST_SHARE_AREA, "backups$", "stage-deleted")
-SMALLEST_PERMISSIBLE_MYSQL_DUMP_FILE_IN_BYTES = 100
-
-ALL_INSTALL_DIRECTORIES = (EPICS_PATH, PYTHON_PATH, PYTHON_3_PATH, GUI_PATH, GUI_PATH_E4, EPICS_UTILS_PATH)
-
-GIGABYTE = 1024 ** 3
-
-RAM_MIN = 7.5 * GIGABYTE  # 8 GB minus a small tolerance.
-RAM_NORMAL_INSTRUMENT = 13 * GIGABYTE  # Should be 14GB ideally, but allow anything over 13GB.
-FREE_DISK_MIN = 30 * GIGABYTE
 
 
 class UpgradeInstrument(object):
@@ -134,7 +33,19 @@ class UpgradeInstrument(object):
             genie_python3_dir: directory to install genie_python 3 from
             file_utils : collection of file utilities
         """
-        self._upgrade_tasks = UpgradeTasks(
+        self._client_tasks = ClientTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._mysql_tasks = MysqlTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._python_tasks = PythonTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._server_tasks = ServerTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._system_tasks = SystemTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._vhd_tasks = VHDTasks(
+            user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
+        self._backup_tasks = BackupTasks(
             user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir, file_utils)
 
     @staticmethod
@@ -150,85 +61,83 @@ class UpgradeInstrument(object):
         """
         Run a complete test upgrade on the current system
         """
-        self._upgrade_tasks.user_confirm_upgrade_type_on_machine('Training Machine')
-        self._upgrade_tasks.install_or_upgrade_git()
-        self._upgrade_tasks.stop_ibex_server()
-        self._upgrade_tasks.remove_old_ibex()
-        self._upgrade_tasks.clean_up_desktop_ibex_training_folder()
-        self._upgrade_tasks.remove_settings()
-        self._upgrade_tasks.install_settings()
-        self._upgrade_tasks.install_ibex_server(self._should_install_utils())
-        self._upgrade_tasks.install_genie_python3()
-        self._upgrade_tasks.install_ibex_client()
-        self._upgrade_tasks.upgrade_notepad_pp()
-        self._upgrade_tasks.change_shortcuts_to_python_3()
+        self._system_tasks.user_confirm_upgrade_type_on_machine('Training Machine')
+        self._system_tasks.install_or_upgrade_git()
+        self._backup_tasks.remove_old_ibex()
+        self._system_tasks.clean_up_desktop_ibex_training_folder()
+        self._server_tasks.remove_settings()
+        self._server_tasks.install_settings()
+        self._server_tasks.install_ibex_server(self._should_install_utils())
+        self._python_tasks.install_genie_python3()
+        self._client_tasks.install_ibex_client()
+        self._system_tasks.upgrade_notepad_pp()
+        self._python_tasks.change_shortcuts_to_python_3()
 
     def remove_all_and_install_client_and_server(self):
         """
         Either install or upgrade the ibex client and server
         """
-        self._upgrade_tasks.confirm(
+        self._system_tasks.confirm(
             "This script removes IBEX client and server and installs the latest build of both, and upgrade the "
             "config/schema without any extra steps. Proceed?")
 
-        self._upgrade_tasks.user_confirm_upgrade_type_on_machine('Client/Server Machine')
-        self._upgrade_tasks.stop_ibex_server()
-        self._upgrade_tasks.remove_old_ibex()
-        self._upgrade_tasks.install_ibex_server(self._should_install_utils())
-        self._upgrade_tasks.install_genie_python3()
-        self._upgrade_tasks.install_e4_ibex_client()
-        self._upgrade_tasks.upgrade_instrument_configuration()
+        self._system_tasks.user_confirm_upgrade_type_on_machine('Client/Server Machine')
+        self._backup_tasks.remove_old_ibex()
+        self._server_tasks.install_ibex_server(self._should_install_utils())
+        self._python_tasks.install_genie_python3()
+        self._client_tasks.install_e4_ibex_client()
+        self._server_tasks.upgrade_instrument_configuration()
 
     def run_instrument_tests(self):
         """
         Run through client and server tests once installation / deployment has completed.
         """
-        self._upgrade_tasks.perform_client_tests()
-        self._upgrade_tasks.perform_server_tests()
-        self._upgrade_tasks.inform_instrument_scientists()
+        self._client_tasks.perform_client_tests()
+        self._server_tasks.perform_server_tests()
+        self._system_tasks.inform_instrument_scientists()
 
     def run_instrument_install(self):
         """
         Do a first installation of IBEX on a new instrument.
         """
-        self._upgrade_tasks.confirm("This script performs a first-time full installation of the IBEX server and client"
+        self._system_tasks.confirm("This script performs a first-time full installation of the IBEX server and client"
                                     " on a new instrument. Proceed?")
 
-        self._upgrade_tasks.check_resources()
-        self._upgrade_tasks.install_or_upgrade_git()
-        self._upgrade_tasks.check_java_installation()
+        self._system_tasks.check_resources()
+        self._system_tasks.install_or_upgrade_git()
+        self._system_tasks.check_java_installation()
 
-        self._upgrade_tasks.remove_seci_shortcuts()
-        self._upgrade_tasks.remove_seci_one()
-        self._upgrade_tasks.remove_treesize_shortcuts()
-        self._upgrade_tasks.restrict_ie()
+        self._system_tasks.remove_seci_shortcuts()
+        self._system_tasks.remove_seci_one()
+        self._system_tasks.remove_treesize_shortcuts()
+        self._system_tasks.restrict_ie()
 
-        self._upgrade_tasks.install_ibex_server(self._should_install_utils())
-        self._upgrade_tasks.install_genie_python3()
-        self._upgrade_tasks.install_mysql()
-        self._upgrade_tasks.install_ibex_client()
-        self._upgrade_tasks.setup_config_repository()
-        self._upgrade_tasks.upgrade_instrument_configuration()
-        self._upgrade_tasks.configure_com_ports()
-        self._upgrade_tasks.setup_calibrations_repository()
-        self._upgrade_tasks.update_calibrations_repository()
-        self._upgrade_tasks.apply_changes_noted_in_release_notes()
-        self._upgrade_tasks.update_release_notes()
-        self._upgrade_tasks.restart_vis()
-        self._upgrade_tasks.install_wiring_tables()
-        self._upgrade_tasks.configure_motion()
-        self._upgrade_tasks.add_nagios_checks()
-        self._upgrade_tasks.update_instlist()
-        self._upgrade_tasks.update_web_dashboard()
-        self._upgrade_tasks.update_kafka_topics()
-        self._upgrade_tasks.put_autostart_script_in_startup_area()
-        self._upgrade_tasks.update_script_definitions()
+        self._server_tasks.install_ibex_server(self._should_install_utils())
+        self._python_tasks.install_genie_python3()
+        self._mysql_tasks.install_mysql()
+        self._client_tasks.install_ibex_client()
+        self._server_tasks.setup_config_repository()
+        self._server_tasks.upgrade_instrument_configuration()
+        self._system_tasks.configure_com_ports()
+        self._server_tasks.setup_calibrations_repository()
+        self._server_tasks.update_calibrations_repository()
+        self._system_tasks.apply_changes_noted_in_release_notes()
+        self._system_tasks.update_release_notes()
+        self._system_tasks.restart_vis()
+        self._server_tasks.install_wiring_tables()
+        self._server_tasks.configure_motion()
+        self._system_tasks.add_nagios_checks()
+        self._system_tasks.update_instlist()
+        self._system_tasks.update_web_dashboard()
+        self._system_tasks.update_kafka_topics()
+        self._system_tasks.put_autostart_script_in_startup_area()
+        self._python_tasks.update_script_definitions()
 
     def run_instrument_deploy(self):
         """
         Deploy a full IBEX upgrade on an existing instrument.
         """
-        self._upgrade_tasks.confirm(
+        self._system_tasks.confirm(
             "This script performs a full upgrade of the IBEX server and client on an existing instrument. Proceed?")
         self.run_instrument_deploy_pre_stop()
         self.run_instrument_deploy_main()
@@ -240,16 +149,16 @@ class UpgradeInstrument(object):
 
             Current the server can not be started in this python script.
         """
-        self._upgrade_tasks.start_ibex_gui()
-        self._upgrade_tasks.restart_vis()
-        self._upgrade_tasks.perform_client_tests()
-        self._upgrade_tasks.perform_server_tests()
-        self._upgrade_tasks.save_motor_parameters_to_file()
-        self._upgrade_tasks.save_blocks_to_file()
-        self._upgrade_tasks.save_blockserver_pv_to_file()
-        self._upgrade_tasks.put_autostart_script_in_startup_area()
-        self._upgrade_tasks.change_shortcuts_to_python_3()
-        self._upgrade_tasks.inform_instrument_scientists()
+        self._client_tasks.start_ibex_gui()
+        self._system_tasks.restart_vis()
+        self._client_tasks.perform_client_tests()
+        self._server_tasks.perform_server_tests()
+        self._server_tasks.save_motor_parameters_to_file()
+        self._server_tasks.save_blocks_to_file()
+        self._server_tasks.save_blockserver_pv_to_file()
+        self._system_tasks.put_autostart_script_in_startup_area()
+        self._python_tasks.change_shortcuts_to_python_3()
+        self._system_tasks.inform_instrument_scientists()
 
     def run_instrument_deploy_main(self):
         """
@@ -257,22 +166,22 @@ class UpgradeInstrument(object):
 
             Current the server can not be started or stopped in this python script.
         """
-        self._upgrade_tasks.install_or_upgrade_git()
-        self._upgrade_tasks.check_java_installation()
-        self._upgrade_tasks.backup_old_directories()
-        self._upgrade_tasks.backup_database()
-        self._upgrade_tasks.truncate_database()
-        self._upgrade_tasks.install_ibex_server(self._should_install_utils())
-        self._upgrade_tasks.install_genie_python3()
-        self._upgrade_tasks.install_mysql()
+        self._system_tasks.install_or_upgrade_git()
+        self._system_tasks.check_java_installation()
+        self._backup_tasks.backup_old_directories()
+        self._mysql_tasks.backup_database()
+        self._mysql_tasks.truncate_database()
+        self._server_tasks.install_ibex_server(self._should_install_utils())
+        self._python_tasks.install_genie_python3()
+        self._mysql_tasks.install_mysql()
 
-        self._upgrade_tasks.install_ibex_client()
-        self._upgrade_tasks.upgrade_instrument_configuration()
-        self._upgrade_tasks.update_calibrations_repository()
-        self._upgrade_tasks.apply_changes_noted_in_release_notes()
-        self._upgrade_tasks.update_release_notes()
-        self._upgrade_tasks.reapply_hotfixes()
-        self._upgrade_tasks.update_script_definitions()
+        self._client_tasks.install_ibex_client()
+        self._server_tasks.upgrade_instrument_configuration()
+        self._server_tasks.update_calibrations_repository()
+        self._system_tasks.apply_changes_noted_in_release_notes()
+        self._system_tasks.update_release_notes()
+        self._system_tasks.reapply_hotfixes()
+        self._python_tasks.update_script_definitions()
 
     def run_instrument_deploy_pre_stop(self):
         """
@@ -280,33 +189,32 @@ class UpgradeInstrument(object):
 
             Current the server can not be started or stopped in this python script.
         """
-        self._upgrade_tasks.user_confirm_upgrade_type_on_machine('Client/Server Machine')
-        self._upgrade_tasks.record_running_vis()
-        self._upgrade_tasks.save_motor_parameters_to_file()
-        self._upgrade_tasks.save_blocks_to_file()
-        self._upgrade_tasks.save_blockserver_pv_to_file()
-        self._upgrade_tasks.stop_ibex_server()
+        self._system_tasks.user_confirm_upgrade_type_on_machine('Client/Server Machine')
+        self._system_tasks.record_running_vis()
+        self._server_tasks.save_motor_parameters_to_file()
+        self._server_tasks.save_blocks_to_file()
+        self._server_tasks.save_blockserver_pv_to_file()
 
     def run_truncate_database(self):
         """
         Backup and truncate databases only
         """
-        self._upgrade_tasks.backup_database()
-        self._upgrade_tasks.truncate_database()
+        self._mysql_tasks.backup_database()
+        self._mysql_tasks.truncate_database()
 
     def run_force_upgrade_mysql(self):
         """:key
          Do upgrade of mysql, with data dump.
         """
-        self._upgrade_tasks.install_mysql(force=True)
+        self._mysql_tasks.install_mysql(force=True)
 
     def run_developer_update(self):
         """
         Update all the developer tools to latest version
         """
-        self._upgrade_tasks.install_mysql(force=False)
-        self._upgrade_tasks.check_java_installation()
-        self._upgrade_tasks.install_or_upgrade_git()
+        self._mysql_tasks.install_mysql(force=False)
+        self._system_tasks.check_java_installation()
+        self._system_tasks.install_or_upgrade_git()
 
     def run_vhd_creation(self):
         """
@@ -315,36 +223,36 @@ class UpgradeInstrument(object):
         Note: this will run under jenkins, don't add interactive tasks to this list.
         """
         # To try to start from a clean state if we already have partially mounted VHDs
-        self._upgrade_tasks.request_dismount_vhds()
+        self._vhd_tasks.request_dismount_vhds()
 
-        self._upgrade_tasks.copy_vhds_to_local_area()
-        self._upgrade_tasks.request_mount_vhds()
+        self._vhd_tasks.copy_vhds_to_local_area()
+        self._vhd_tasks.request_mount_vhds()
         try:
-            self._upgrade_tasks.install_ibex_server(True)
-            self._upgrade_tasks.install_genie_python3()
-            self._upgrade_tasks.install_mysql_for_vhd()
-            self._upgrade_tasks.install_e4_ibex_client()
-            self._upgrade_tasks.setup_config_repository()
-            self._upgrade_tasks.upgrade_instrument_configuration()
-            self._upgrade_tasks.setup_calibrations_repository()
-            self._upgrade_tasks.update_calibrations_repository()
-            self._upgrade_tasks.initialize_var_dir()
+            self._server_tasks.install_ibex_server(True)
+            self._python_tasks.install_genie_python3()
+            self._mysql_tasks.install_mysql_for_vhd()
+            self._client_tasks.install_e4_ibex_client()
+            self._server_tasks.setup_config_repository()
+            self._server_tasks.upgrade_instrument_configuration()
+            self._server_tasks.setup_calibrations_repository()
+            self._server_tasks.update_calibrations_repository()
+            self._vhd_tasks.initialize_var_dir()
         finally:
-            self._upgrade_tasks.request_dismount_vhds()
+            self._vhd_tasks.request_dismount_vhds()
 
-        self._upgrade_tasks.deploy_vhds()
+        self._vhd_tasks.deploy_vhds()
 
     def mount_vhds(self):
         """
         Task which actually mounts the VHDs (will be run as admin)
         """
-        self._upgrade_tasks.mount_vhds()
+        self._vhd_tasks.mount_vhds()
 
     def dismount_vhds(self):
         """
         Task which actually dismounts the VHDs (will be run as admin)
         """
-        self._upgrade_tasks.dismount_vhds()
+        self._vhd_tasks.dismount_vhds()
 
 
 # All possible upgrade tasks
@@ -392,1257 +300,3 @@ UPGRADE_TYPES = {
         UpgradeInstrument.dismount_vhds,
         "task to dismount VHDs if needed"),
 }
-
-
-class UpgradeTasks(object):
-    """
-    Class containing separate upgrade tasks.
-    """
-
-    def __init__(self, user_prompt, server_source_dir, client_source_dir, client_e4_source_dir, genie_python3_dir,
-                 file_utils=FileUtils()):
-        """
-        Initializer.
-        Args:
-            user_prompt (ibex_install_utils.user_prompt.UserPrompt): a object to allow prompting of the user
-            server_source_dir: directory to install ibex server from
-            client_source_dir: directory to install ibex client from
-            client_e4_source_dir: directory to install ibex E4 client from
-            genie_python3_dir: directory to install genie python from
-            file_utils : collection of file utilities
-        """
-        self.prompt = user_prompt  # This is needed to allow @tasks to work
-        self._server_source_dir = server_source_dir
-        self._client_source_dir = client_source_dir
-        self._client_e4_source_dir = client_e4_source_dir
-        self._genie_python_3_source_dir = genie_python3_dir
-        self._file_utils = file_utils
-
-        self._machine_name = self._get_machine_name()
-
-        self._ca = CaWrapper()
-
-    @staticmethod
-    def _get_machine_name():
-        """
-        Returns:
-            The current machine name
-
-        """
-        return socket.gethostname()
-
-    @staticmethod
-    def _get_instrument_name():
-        """
-        Returns:
-            The name of the current instrument
-        """
-        return UpgradeTasks._get_machine_name().replace("NDX", "")
-
-    @staticmethod
-    def _get_config_path():
-
-        """
-        Returns:
-            The path to the instrument's configurations directory
-
-        """
-        return os.path.join(INSTRUMENT_BASE_DIR, SETTINGS_CONFIG_FOLDER, UpgradeTasks._get_machine_name())
-
-    def confirm(self, message):
-        """
-        Ask user to confirm correct script was chosen.
-        """
-        self.prompt.prompt_and_raise_if_not_yes(message, default="Y")
-
-    def stop_ibex_server(self):
-        """
-        Stop the current IBEX server running. Current this can not be run because it kills any python
-        processes including this one.
-
-        Returns:
-
-        """
-        # with Task("Stopping IBEX server", self.prompt):
-        #     RunProcess(EPICS_PATH, "stop_ibex_server.bat").run()
-        pass
-
-    @task("Removing old version of IBEX")
-    def remove_old_ibex(self):
-        """
-        Removes older versions of IBEX server, client, genie_python and epics utils
-        Returns:
-
-        """
-
-        for path in ALL_INSTALL_DIRECTORIES:
-            self._file_utils.remove_tree(path, self.prompt)
-
-    @task("Removing training folder on desktop ...")
-    def clean_up_desktop_ibex_training_folder(self):
-        """
-        Remove training folder from the desktop
-        Returns:
-
-        """
-        self._file_utils.remove_tree(DESKTOP_TRAINING_FOLDER_PATH, self.prompt)
-
-    @task("Removing old settings file")
-    def remove_settings(self):
-        """
-        remove old settings
-        Returns:
-
-        """
-        self._file_utils.remove_tree(SETTINGS_CONFIG_PATH, self.prompt)
-
-    @task("Install settings")
-    def install_settings(self):
-        """
-        Install new settings from the current folder
-        Returns:
-
-        """
-
-        self._file_utils.mkdir_recursive(SETTINGS_CONFIG_PATH)
-        settings_path = os.path.join(SETTINGS_CONFIG_PATH, self._machine_name)
-
-        shutil.copytree(SOURCE_MACHINE_SETTINGS_CONFIG_PATH, settings_path)
-
-        inst_name = self._machine_name.lower()
-        for p in ["ndx", "nde"]:
-            if inst_name.startswith(p):
-                inst_name = inst_name.lower()[len(p):]
-                break
-        inst_name = inst_name.replace("-", "_")
-        self._file_utils.move_file(
-            os.path.join(settings_path, "Python", "init_inst_name.py"),
-            os.path.join(settings_path, "Python", "init_{inst_name}.py".format(inst_name=inst_name)),
-            self.prompt)
-
-        shutil.copytree(SOURCE_MACHINE_SETTINGS_COMMON_PATH, os.path.join(SETTINGS_CONFIG_PATH, "common"))
-
-    @task("Record running vis")
-    def record_running_vis(self):
-        """
-        Get user to record running vis
-        """
-        self.prompt.prompt_and_raise_if_not_yes("Write down any VIs which are running for use later?")
-
-    @task("Upgrading Notepad++. Please follow system dialogs")
-    def upgrade_notepad_pp(self):
-        """
-        Install (start installation of) notepad ++
-        Returns:
-
-        """
-        RunProcess(working_dir=APPS_BASE_DIR,
-                   executable_file="GUP.exe",
-                   executable_directory=os.path.join("C:\\", "Program Files (x86)", "Notepad++", "updater")).run()
-
-    @task("Installing IBEX Server")
-    def install_ibex_server(self, with_utils):
-        """
-        Install ibex server.
-        Args:
-            with_utils: True also install epics utils using icp binaries; False don't
-
-        """
-        self._file_utils.mkdir_recursive(APPS_BASE_DIR)
-        RunProcess(self._server_source_dir, "install_to_inst.bat", prog_args=["NOLOG"]).run()
-        if with_utils and self.prompt.confirm_step("install icp binaries"):
-            RunProcess(EPICS_PATH, "create_icp_binaries.bat").run()
-
-    @task("Installing Genie Python 3")
-    def install_genie_python3(self):
-        """
-        Install ibex server.
-        """
-        self._file_utils.mkdir_recursive(APPS_BASE_DIR)
-        RunProcess(self._genie_python_3_source_dir, "genie_python_install.bat").run()
-
-    @task("Installing IBEX Client")
-    def install_ibex_client(self):
-        """
-        Install the ibex client (which also installs genie python).
-
-        """
-        self._install_set_version_of_ibex_client(self._client_source_dir)
-
-    @task("Installing IBEX Client")
-    def install_e4_ibex_client(self):
-        """
-        Install the ibex client E4 version (which also installs genie python).
-
-        """
-        source_dir = self._client_e4_source_dir
-        if source_dir is None:
-            self.prompt.prompt_and_raise_if_not_yes("The E4 client path has not been set; continue with installation?")
-        else:
-            self._install_set_version_of_ibex_client(source_dir)
-
-    def _install_set_version_of_ibex_client(self, source_dir):
-        """
-        Install a given version of the Ibex client.
-        Args:
-            source_dir: source directory for the client
-        """
-        self._file_utils.mkdir_recursive(APPS_BASE_DIR)
-
-        RunProcess(source_dir, "install_client.bat", press_any_key=True).run()
-
-    @task("Starting IBEX gui")
-    def start_ibex_gui(self):
-        """
-        Start the IBEX GUI
-        :return:
-        """
-        subprocess.Popen([os.path.join(GUI_PATH_E4, "ibex-client.exe")], cwd=GUI_PATH_E4)
-
-    def user_confirm_upgrade_type_on_machine(self, machine_type):
-        """
-        Print information about the current upgrade and prompt the user
-        Returns:
-        Raises UserStop: when the user doesn't want to continue
-
-        """
-        print("Upgrade {0} as a {1}".format(self._machine_name, machine_type))
-        print("    Server source: {0}".format(self._server_source_dir))
-        print("    Client source: {0}".format(self._client_source_dir))
-        print("    Python 3 source: {0}".format(self._genie_python_3_source_dir))
-        answer = self.prompt.prompt("Continue? [Y/N]", ["Y", "N"], "Y")
-        if answer != "Y":
-            raise UserStop()
-
-    @task("Set up configuration repository")
-    def setup_config_repository(self):
-        """
-        Creates the configuration repository, and swaps or creates a branch for the instrument.
-
-        """
-        inst_name = self._machine_name
-
-        subprocess.call("git config --global core.autocrlf true")
-        subprocess.call("git config --global credential.helper wincred")
-        subprocess.call("git config --global user.name spudulike")
-        set_user_email = "git config --global user.email spudulike@{}.isis.cclrc.ac.uk"
-        subprocess.call(set_user_email.format(inst_name.lower()))
-
-        if not os.path.exists(SETTINGS_CONFIG_PATH):
-            os.makedirs(SETTINGS_CONFIG_PATH)
-
-        subprocess.call(
-            "git clone http://spudulike@control-svcs.isis.cclrc.ac.uk/gitroot/instconfigs/inst.git {}".format(
-                inst_name), cwd=SETTINGS_CONFIG_PATH)
-
-        inst_config_path = os.path.join(SETTINGS_CONFIG_PATH, inst_name)
-        subprocess.call("git pull", cwd=inst_config_path)
-
-        branch_exists = subprocess.call("git checkout {}".format(inst_name), cwd=inst_config_path) == 0
-        if not branch_exists:
-            subprocess.call("git checkout -b {}".format(inst_name), cwd=inst_config_path)
-
-        inst_scripts_path = os.path.join(inst_config_path, "Python")
-        if not os.path.exists(os.path.join(inst_scripts_path, "init_{}.py".format(inst_name.lower()))):
-            try:
-                os.rename(os.path.join(inst_scripts_path, "init_inst_name.py"),
-                          os.path.join(inst_scripts_path, "init_{}.py".format(inst_name.lower())))
-                subprocess.call("git add init_{}.py".format(inst_name.lower()), cwd=inst_scripts_path)
-                subprocess.call("git rm init_inst_name.py", cwd=inst_scripts_path)
-                subprocess.call('git commit -m"create initial python"'.format(inst_name), cwd=inst_config_path)
-                subprocess.call("git push --set-upstream origin {}".format(inst_name), cwd=inst_config_path)
-            except Exception as e:
-                self.prompt.prompt_and_raise_if_not_yes(
-                    "Something went wrong setting up the configurations repository. Please resolve manually, "
-                    "instructions are in the developers manual under "
-                    "First-time-installing-and-building-(Windows): \n {}".format(e.message))
-
-    @task("Upgrading instrument configuration")
-    def upgrade_instrument_configuration(self):
-        """
-        Update the configuration on the instrument using its upgrade config script.
-        """
-        manual_prompt = "Merge the master configurations branch into the instrument configuration. "\
-                        "From C:\Instrument\Settings\config\[machine name] run:\n"\
-                        "    0. Clean up any in progress merge (e.g. git merge --abort)\n"\
-                        "    1. git checkout master\n"\
-                        "    2. git pull\n"\
-                        "    3. git checkout [machine name]\n"\
-                        "    4. git merge master\n"\
-                        "    5. Resolve any merge conflicts\n"\
-                        "    6. git push\n"
-        automatic_prompt = "Attempt automatic configuration merge?"
-        if self.prompt.confirm_step(automatic_prompt):
-            try:
-                repo = git.Repo(os.path.join(SETTINGS_CONFIG_PATH, self._machine_name))
-                if repo.active_branch.name != self._machine_name:
-                    print("Git branch, '{}', is not the same as machine name ,'{}' ".format(
-                        repo.active_branch, self._machine_name))
-                    raise ErrorInTask("Git branch is not the same as machine name")
-                else:
-                    print("     fetch: {}".format(repo.git.fetch()))
-                    print("     merge: {}".format(repo.git.merge("origin/master")))
-                    # no longer push let the instrument do that on start up if needed
-            except git.GitCommandError as e:
-                print("Error doing automatic merge, please perform the merge manually: {}".format(e))
-                self.prompt.prompt_and_raise_if_not_yes(manual_prompt)
-        else:
-            self.prompt.prompt_and_raise_if_not_yes(manual_prompt)
-
-        RunProcess(CONFIG_UPGRADE_SCRIPT_DIR, "upgrade.bat", capture_pipes=False).run()
-
-    @task("Remove SECI shortcuts")
-    def remove_seci_shortcuts(self):
-        """
-        Remove (or at least ask the user to remove) all Seci shortcuts
-        """
-        for path in AUTOSTART_LOCATIONS:
-            if os.path.exists(path):
-                self.prompt.prompt_and_raise_if_not_yes(
-                    "SECI autostart found in {}, delete this.".format(path))
-
-        self.prompt.prompt_and_raise_if_not_yes("Remove task bar shortcut to SECI")
-        self.prompt.prompt_and_raise_if_not_yes("Remove desktop shortcut to SECI")
-        self.prompt.prompt_and_raise_if_not_yes("Remove start menu shortcut to SECI")
-
-    @task("Update script generator script definitions")
-    def update_script_definitions(self):
-        """
-        Update (or at least ask the user to update) the script definitions used by the script generator.
-        """
-        if os.path.exists("C:\\ScriptGeneratorConfigs") or os.path.exists("C:\\ScriptDefinitions"):
-            self.prompt.prompt("Update the script definitions for the script generator (likely in C:\\ScriptDefinitions or C:\\ScriptGeneratorConfigs)." + \
-                "Check with the scientists that it is ok to do this." + \
-                "You can do it by git pull, you may need to merge changes made on the instrument.")
-
-    @task("Remove Treesize shortcuts")
-    def remove_treesize_shortcuts(self):
-        """
-        Remove (or at least ask the user to remove) all Treesize shortcuts.
-
-        For justification see https://github.com/ISISComputingGroup/IBEX/issues/4214
-        """
-        self.prompt.prompt_and_raise_if_not_yes("Remove task bar shortcut to Treesize if it exists")
-        self.prompt.prompt_and_raise_if_not_yes("Remove desktop shortcut to Treesize if it exists")
-        self.prompt.prompt_and_raise_if_not_yes("Remove start menu shortcut to Treesize if it exists")
-
-    @task("Remove SECI 1 Path")
-    def remove_seci_one(self):
-        """
-        Removes SECI 1
-        """
-        if os.path.exists(SECI_ONE_PATH):
-            try:
-                self._file_utils.remove_tree(SECI_ONE_PATH, self.prompt, use_robocopy=False)
-            except (IOError, WindowsError) as e:
-                self.prompt.prompt_and_raise_if_not_yes("Failed to remove SECI 1 (located in '{}') because "
-                                                        "'{}'. Please remove it manually and type 'Y' to "
-                                                        "confirm".format(SECI_ONE_PATH, e.message))
-
-    @task("Set up calibrations repository")
-    def setup_calibrations_repository(self):
-        """
-        Set up the calibration repository
-        """
-        if os.path.isdir(CALIBRATION_PATH):
-            if self.prompt.prompt("Calibrations directory already exists. Update calibrations repository?",
-                                  ["Y", "N"], "N") == "Y":
-                self.update_calibrations_repository()
-        else:
-            exit_code = subprocess.call("git clone http://control-svcs.isis.cclrc.ac.uk/gitroot/instconfigs/common.git "
-                                        "C:\Instrument\Settings\config\common")
-            if exit_code is not 0:
-                raise ErrorInRun("Failed to set up common calibration directory.")
-
-    @task("Updating calibrations repository")
-    def update_calibrations_repository(self):
-        """
-        Update the calibration repository
-        """
-        try:
-            repo = git.Repo(CALIBRATION_PATH)
-            repo.git.pull()
-        except git.GitCommandError:
-            self.prompt.prompt_and_raise_if_not_yes("There was an error pulling the calibrations repo.\n"
-                                                    "Manually pull it. Path='{}'".format(CALIBRATION_PATH))
-
-    @task("Install java")
-    def check_java_installation(self):
-        """
-        Checks Java installation
-        """
-
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Upgrade openJDK installation by following:\r\n"
-            "https://github.com/ISISComputingGroup/ibex_developers_manual/wiki/Upgrade-Java\r\n\r\n"
-            "After following the installer, ensure you close and then re-open your remote desktop session (This "
-            "is a workaround for windows not immediately picking up new environment variables)")
-
-    @task("Configure COM ports")
-    def configure_com_ports(self):
-        """
-        Configure the COM ports
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Using NPort Administrator (available under /Kits$/CompGroup/Utilities/), check that the COM ports "
-            "on this machine are configured to standard, i.e.:\n"
-            "- Moxa 1 starts at COM5\n"
-            "- Moxa 2 starts at COM21\n"
-            "- etc.\n")
-
-    @staticmethod
-    def _get_backup_dir():
-        """
-        Returns: The backup dir, will create it if needed (both old and dir).
-        Raises: IOError if the base dir doesn't exist
-        """
-        new_backup_dir = os.path.join(BACKUP_DIR, "ibex_backup_{}".format(UpgradeTasks._today_date_for_filenames()))
-
-        if not os.path.exists(BACKUP_DATA_DIR):
-            raise IOError("Base directory does not exist {} should be a provided linked dir".format(BACKUP_DATA_DIR))
-        if not os.path.exists(BACKUP_DIR):
-            os.mkdir(BACKUP_DIR)
-        if not os.path.exists(new_backup_dir):
-            os.mkdir(new_backup_dir)
-        return new_backup_dir
-
-    @staticmethod
-    def _today_date_for_filenames():
-        return date.today().strftime("%Y_%m_%d")
-
-    def _backup_dir(self, src, copy=True):
-        backup_dir = os.path.join(self._get_backup_dir(), os.path.basename(src))
-        if src in os.getcwd():
-            self.prompt.prompt_and_raise_if_not_yes(
-                "You appear to be trying to delete the folder, {}, containing the current working directory {}. "
-                "Please do this manually to be on the safe side".format(src, os.getcwd()))
-        elif os.path.exists(backup_dir):
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Backup dir {} already exist. Please backup this app manually".format(backup_dir))
-        elif os.path.exists(src):
-            if copy:
-                print("Copying {} to {}".format(src, backup_dir))
-                shutil.copytree(src, backup_dir)
-            else:
-                print("Moving {} to {}".format(src, backup_dir))
-                self._file_utils.move_dir(src, backup_dir, self.prompt)
-
-    @task("Backup old directories")
-    def backup_old_directories(self):
-        """
-        Backup old directories
-        """
-        if os.path.exists(BACKUP_DATA_DIR):
-            if not os.path.exists(BACKUP_DIR):
-                os.mkdir(BACKUP_DIR)
-
-            # Delete all but the oldest backup
-            current_backups = [os.path.join(BACKUP_DIR, d) for d in os.listdir(BACKUP_DIR)
-                               if os.path.isdir(os.path.join(BACKUP_DIR, d)) and d.startswith("ibex_backup")]
-            if len(current_backups) > 0:
-                all_but_newest_backup = sorted(current_backups, key=os.path.getmtime)[:-1]
-                backups_to_delete = all_but_newest_backup
-            else:
-                backups_to_delete = []
-
-            for d in backups_to_delete:
-                print("Removing backup {}".format(d))
-                self._file_utils.remove_tree(os.path.join(BACKUP_DIR, d), self.prompt)
-
-            # Move the folders
-            for app_path in ALL_INSTALL_DIRECTORIES:
-                self._backup_dir(app_path, copy=False)
-
-            # Backup settings and autosave
-            self._backup_dir(os.path.join(INSTRUMENT_BASE_DIR, "Settings"))
-            self._backup_dir(os.path.join(INSTRUMENT_BASE_DIR, "var", "Autosave"))
-        else:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Unable to find data directory '{}'. Please backup the current installation of IBEX "
-                "manually".format(BACKUP_DATA_DIR))
-
-    def _get_mysql_dir(self):
-        """
-        Returns the mysql 8 default install directory if it exists, else 5.7.
-
-        """
-        if os.path.exists(MYSQL8_INSTALL_DIR):
-            mysql_bin_dir = os.path.join(MYSQL8_INSTALL_DIR, "bin")
-        else:
-            mysql_bin_dir = os.path.join(MYSQL57_INSTALL_DIR, "bin")
-
-        return mysql_bin_dir
-
-    @task("Backup database")
-    def backup_database(self):
-        """
-        Backup the database
-        """
-        result_file = os.path.join(self._get_backup_dir(),
-                                   SQLDUMP_FILE_TEMPLATE.format(UpgradeTasks._today_date_for_filenames()))
-
-        mysql_bin_dir = self._get_mysql_dir()
-
-        dump_command = ["-u", "root", "-p", "--all-databases", "--single-transaction",
-                        "--result-file={}".format(result_file)]
-        RunProcess(MYSQL_FILES_DIR, "mysqldump.exe", executable_directory=mysql_bin_dir,
-                   prog_args=dump_command,
-                   capture_pipes=False).run()
-
-        if os.path.getsize(result_file) < SMALLEST_PERMISSIBLE_MYSQL_DUMP_FILE_IN_BYTES:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Dump file '{}' seems to be small is it correct? ".format(result_file))
-
-        self._file_utils.move_file(result_file, os.path.join(STAGE_DELETED, self._get_machine_name()),
-                                   self.prompt)
-
-    def _backup_data(self):
-        """
-        Backup the data for transfer. This dumps just the data not the schema.
-        """
-        result_file = os.path.join(self._get_backup_dir(),
-                                   SQLDUMP_FILE_TEMPLATE.format(UpgradeTasks._today_date_for_filenames()))
-
-        mysql_bin_dir = self._get_mysql_dir()
-        dump_command = ["-u", "root", "-p", "--single-transaction",
-                        "--result-file={}".format(result_file), "--no-create-db", "--no-create-info", "--skip-triggers",
-                        "--databases", "alarm", "archive", "exp_data", "iocdb", "journal", "msg_log"]
-        RunProcess(MYSQL_FILES_DIR, "mysqldump.exe", executable_directory=mysql_bin_dir,
-                   prog_args=dump_command,
-                   capture_pipes=False).run()
-
-        if os.path.getsize(result_file) < SMALLEST_PERMISSIBLE_MYSQL_DUMP_FILE_IN_BYTES:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Dump file '{}' seems to be small is it correct? ".format(result_file))
-
-    def _reload_backup_data(self):
-        """
-        Reload backup the data
-        """
-        result_file = os.path.join(self._get_backup_dir(),
-                                   SQLDUMP_FILE_TEMPLATE.format(UpgradeTasks._today_date_for_filenames()))
-
-        mysql_bin_dir = self._get_mysql_dir()
-        read_dump_command = ["-u", "root", "-p", "--force"]
-        RunProcess(MYSQL_FILES_DIR, "mysql.exe", executable_directory=mysql_bin_dir,
-                   prog_args=read_dump_command,
-                   capture_pipes=False, std_in=open(result_file)).run()
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Check that there are only 16 errors from mysql. These are 1062 error fail to insert primary key. "
-            "These are for constants added by the creation script, e.g. archive severity.")
-
-    @task("Truncate database")
-    def truncate_database(self):
-        """
-        Truncate the message log, sample and alarms tables
-        """
-        try:
-            mysql_bin_dir = self._get_mysql_dir()
-
-            sql_command = "truncate table msg_log.message;truncate table archive.sample;truncate table alarm.pv"
-
-            RunProcess(MYSQL_FILES_DIR, "mysql.exe", executable_directory=mysql_bin_dir,
-                       prog_args=["-u", "root", "-p",
-                                  "--execute", sql_command],
-                       capture_pipes=False).run()
-
-        except ErrorInRun as ex:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Unable to run mysql command, please truncate the database manually. "
-                "Error is {}".format(ex.message))
-
-    @task("Update release notes")
-    def update_release_notes(self):
-        """
-        Update the release notes.
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Have you updated the instrument release notes at https://github.com/ISISComputingGroup/IBEX/wiki?")
-
-    def _configure_mysql(self):
-        """
-        Copy mysql settings and run the MySQL configuration script
-        """
-        my_ini_file = os.path.join(EPICS_PATH, "systemsetup", "my.ini")
-        try:
-            shutil.copy(my_ini_file, MYSQL8_INSTALL_DIR)
-        except (OSError, IOError) as e:
-            self.prompt.prompt_and_raise_if_not_yes("Couldn't copy my.ini from {} to {} because {}. "
-                                                    "Please do this manually confirm when complete."
-                                                    .format(my_ini_file, MYSQL8_INSTALL_DIR, e))
-
-        # Restart to pick up new my.ini
-        admin_commands = AdminCommandBuilder()
-        admin_commands.add_command("sc", "stop MYSQL80", expected_return_val=None)
-        # Sleep to wait for service to stop so we can restart it.
-        admin_commands.add_command("ping", "-n 10 127.0.0.1 >nul", expected_return_val=None)
-        admin_commands.add_command("sc", "start MYSQL80", expected_return_val=None)
-        admin_commands.run_all()
-
-    def _remove_old_versions_of_mysql8(self, clean_install):
-        if clean_install:
-            self.prompt.prompt_and_raise_if_not_yes("Warning: this will erase all data held in the MySQL database. "
-                                                    "Are you sure you want to continue?")
-
-        admin_commands = AdminCommandBuilder()
-        admin_commands.add_command("sc", "stop MYSQL80", expected_return_val=None)
-        admin_commands.add_command("sc", "delete MYSQL80", expected_return_val=None)
-        admin_commands.run_all()
-
-        sleep(5)  # Time for service to stop
-
-        if os.path.exists(MYSQL8_INSTALL_DIR):
-            shutil.rmtree(MYSQL8_INSTALL_DIR)
-
-        if clean_install:
-            self._remove_old_mysql_data_dir()
-
-    def _remove_old_mysql_data_dir(self):
-        if os.path.exists(MYSQL_FILES_DIR):
-            shutil.rmtree(MYSQL_FILES_DIR)
-
-    def _create_mysql_binaries(self):
-        os.makedirs(MYSQL8_INSTALL_DIR)
-
-        mysql_unzip_temp = os.path.join(APPS_BASE_DIR, "temp-mysql-unzip")
-
-        with closing(zipfile.ZipFile(MYSQL_ZIP)) as f:
-            f.extractall(mysql_unzip_temp)
-
-        mysql_unzip_temp_release = os.path.join(mysql_unzip_temp, "mysql-{}-winx64".format(MYSQL_LATEST_VERSION))
-        for item in os.listdir(mysql_unzip_temp_release):
-            shutil.move(os.path.join(mysql_unzip_temp_release, item), MYSQL8_INSTALL_DIR)
-
-        shutil.rmtree(mysql_unzip_temp)
-
-    def _initialize_mysql_data_area_for_vhd(self):
-        os.makedirs(os.path.join(MYSQL_FILES_DIR, "data"))
-
-        RunProcess(
-            working_dir=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-            executable_file="mysqld.exe",
-            executable_directory=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-            prog_args=[
-                '--datadir={}'.format(os.path.join(MYSQL_FILES_DIR, "data")),
-                '--initialize-insecure',
-                '--console',
-                '--log-error-verbosity=3'
-            ]
-        ).run()
-
-    def _setup_database_users_and_tables(self, vhd_install=True):
-        mysqld = os.path.join(MYSQL8_INSTALL_DIR, "bin", "mysqld.exe")
-
-        sql_password = self.prompt.prompt("Enter the MySQL root password:", UserPrompt.ANY,
-                                          os.getenv("MYSQL_PASSWORD", "environment variable not set"),
-                                          show_automatic_answer=False)
-
-        if vhd_install:  # Service won't have been started yet
-            # spawn service in background
-            subprocess.Popen(mysqld, creationflags=DETACHED_PROCESS)
-
-            sleep(5)  # Chance for the service to spawn
-
-        RunProcess(
-            executable_directory=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-            working_dir=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-            executable_file="mysql.exe",
-            prog_args=[
-                '-u',
-                'root',
-                '-e',
-                'ALTER USER \'root\'@\'localhost\' IDENTIFIED WITH mysql_native_password BY \'{}\';FLUSH privileges;'
-                    .format(sql_password),
-
-            ],
-            log_command_args=False,  # To make sure password doesn't appear in jenkins log.
-        ).run()
-
-        RunProcess(
-            working_dir=SYSTEM_SETUP_PATH,
-            executable_directory=SYSTEM_SETUP_PATH,
-            executable_file="config_mysql.bat",
-            prog_args=[sql_password],
-            log_command_args=False,  # To make sure password doesn't appear in jenkins log.
-        ).run()
-
-        if vhd_install:
-            # For VHD install only, stop mysql when done
-            RunProcess(
-                executable_directory=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-                working_dir=os.path.join(MYSQL8_INSTALL_DIR, "bin"),
-                executable_file="mysqladmin.exe",
-                prog_args=[
-                    "-u",
-                    "root",
-                    "--password={}".format(sql_password),
-                    "shutdown",
-                ],
-                log_command_args=False,  # To make sure password doesn't appear in jenkins log.
-            ).run()
-
-    def _install_latest_mysql8(self, clean_install):
-        """
-        Install the latest mysql. If this is a clean install remove old data directories first and create a new
-        database
-        Args:
-            clean_install: True to destroy and recreate data directories
-        """
-        self._create_mysql_binaries()
-
-        mysqld = os.path.join(MYSQL8_INSTALL_DIR, "bin", "mysqld.exe")
-
-        admin_commands = AdminCommandBuilder()
-        if clean_install:
-            os.makedirs(MYSQL_FILES_DIR)
-
-            admin_commands.add_command(mysqld, '--datadir="{}" --initialize-insecure --console --log-error-verbosity=3'
-                                   .format(os.path.join(MYSQL_FILES_DIR, "data")))
-
-        # Wait for initialize since admin runner can't wait for completion.
-        # Maybe we can detect completion another way?
-        admin_commands.add_command(mysqld, '--install MYSQL80 --datadir="{}"'
-                                   .format(os.path.join(MYSQL_FILES_DIR, "data")))
-
-        admin_commands.add_command("sc", "start MYSQL80", expected_return_val=None)
-        admin_commands.add_command("sc", "config MYSQL80 start= auto")
-        admin_commands.add_command("sc", "failure MYSQL80 reset= 900 actions= restart/10000/restart/30000/restart/60000")
-        admin_commands.add_command("sc", "failureflag MYSQL80 1")
-        admin_commands.add_command("netsh", "advfirewall firewall delete rule name=mysqld.exe", None)  # remove old firewall rules
-        admin_commands.add_command("netsh", "advfirewall firewall delete rule name=MYSQL8", None)   # remove old firewall rules
-        admin_commands.add_command("netsh", r"advfirewall firewall add rule name=MYSQL8 dir=in action=allow "
-                                            r"program=C:\Instrument\Apps\MySQL\Bin\mysqld.exe enable=yes")
-        
-        admin_commands.run_all()
-
-        sleep(5)  # Time for service to start
-
-        if clean_install:
-            self._setup_database_users_and_tables(vhd_install=False)
-
-    @task("Install latest MySQL for VHD deployment")
-    def install_mysql_for_vhd(self):
-        """
-        Installs MySQL for the VHD creation build.
-
-        Ensure we start from a clean slate. We are creating VHDs so we can assume that no files should exist in
-        C:\instrument\apps\mysql or c:\instrument\var\mysql and delete them if they do exist. This facilitates
-        developer testing/resuming the script if it failed halfway through
-        """
-        for path in [MYSQL_FILES_DIR, MYSQL8_INSTALL_DIR]:
-            if os.path.exists(path):
-                shutil.rmtree(path)
-
-        self._create_mysql_binaries()
-        self._initialize_mysql_data_area_for_vhd()
-
-        my_ini_file = os.path.join(EPICS_PATH, "systemsetup", "my.ini")
-        try:
-            shutil.copy(my_ini_file, MYSQL8_INSTALL_DIR)
-        except (OSError, IOError) as e:
-            self.prompt.prompt_and_raise_if_not_yes("Couldn't copy my.ini from {} to {} because {}. "
-                                                    "Please do this manually confirm when complete."
-                                                    .format(my_ini_file, MYSQL8_INSTALL_DIR, e))
-
-        self._setup_database_users_and_tables(vhd_install=True)
-
-    def _install_vcruntime140(self):
-        if not os.path.exists(VCRUNTIME140):
-            self.prompt.prompt_and_raise_if_not_yes(
-                "MySQL server 8 requires microsoft visual C++ runtime to be installed.\r\n"
-                "Install it from {} and confirm when complete".format(VCRUNTIME140_INSTALLER))
-
-    @task("Install latest MySQL")
-    def install_mysql(self, force=False):
-        """
-        Install mysql and the ibex database schemas
-        Args:
-            force: True delete old data and update
-        """
-        backup_data = False
-        clean_install = True
-        if os.path.exists(os.path.join(MYSQL57_INSTALL_DIR, "bin", "mysql.exe")):
-            self._backup_data()
-            backup_data = True
-            self.prompt.prompt_and_raise_if_not_yes("MySQL 5.7 detected. Please use the MySQL installer application"
-                                                    "to remove MySQL 5.7. When it asks you whether to remove data"
-                                                    "directories, answer yes. Type 'Y' when complete.")
-
-        mysql_8_exe = os.path.join(MYSQL8_INSTALL_DIR, "bin", "mysql.exe")
-
-        if os.path.exists(mysql_8_exe):
-            version = subprocess.check_output("{} --version".format(mysql_8_exe))
-            if MYSQL_LATEST_VERSION in version and not force:
-                answer = self.prompt.prompt("MySQL already appears to be on the latest version ({}) - would you like to"
-                                            " force a reinstall anyway? [Y/N]".format(MYSQL_LATEST_VERSION),
-                                            possibles=["Y", "N"], default="N")
-                if answer == "Y":
-                    force = True
-                else:
-                    return
-            clean_install = force
-            self._remove_old_versions_of_mysql8(clean_install=clean_install)
-
-        self._install_vcruntime140()
-        self._install_latest_mysql8(clean_install=clean_install)
-        self._configure_mysql()
-        if backup_data:
-            self._reload_backup_data()
-
-    @task("Reapply Hotfixes")
-    def reapply_hotfixes(self):
-        """
-        Reapply any hotfixes to the build.
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Have you applied any hotfixes listed that are not fixed by the release, as on the instrument "
-            "release notes at https://github.com/ISISComputingGroup/IBEX/wiki?")
-        if UpgradeTasks._get_machine_name() == "NDEMUONFE":
-            self.prompt.prompt_and_raise_if_not_yes(
-                "MUONFE requires the fix specified in https://github.com/ISISComputingGroup/IBEX/issues/5164, "
-                "which should be specified in the hotfixes, did you apply this?")
-
-    @task("Restart VIs")
-    def restart_vis(self):
-        """
-        Restart Vis which were running on upgrade start.
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Please restart any VIs that were running at the start of the upgrade")
-
-    @task("Client release tests")
-    def perform_client_tests(self):
-        """
-        Test that the client works
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Check that the version displayed in the client is as expected after the deployment")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Confirm that genie_python works from within the client and via genie_python.bat (this includes"
-            "verifying that the 'g.' and 'inst.' prefixes work as expected)"
-            "If the font cannot be seen in the genie_python.bat change default terminal colours to white on black.")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Verify that the current configuration is consistent with the system prior to upgrade")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Verify that all the links from the 'Weblinks' perspective still work (i.e. the address gets resolved)")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Verify that the dashboard gives the instrument name with no NDX prefix (if it does switch to the current instrument)")
-
-    @task("Server release tests")
-    def perform_server_tests(self):
-        """
-        Test that the server works
-        """
-        server_release_tests_url = "https://github.com/ISISComputingGroup/ibex_developers_manual/wiki/" \
-                                   "Server-Release-Tests"
-
-        print("For further details, see {}".format(server_release_tests_url))
-        self.prompt.prompt_and_raise_if_not_yes("Check that blocks are logging as expected")
-
-        print("Checking that configurations are being pushed to the appropriate repository")
-        repo = git.Repo(self._get_config_path())
-        repo.git.fetch()
-        status = repo.git.status()
-        print("Current repository status is: {}".format(status))
-        if "up-to-date with 'origin/{}".format(self._get_machine_name()) in status:
-            print("Configurations updating correctly")
-        else:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Unexpected git status. Please confirm that configurations are being pushed to the appropriate "
-                "remote repository")
-
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Check that the web dashboard for this instrument is updating "
-            "correctly: http://dataweb.isis.rl.ac.uk/IbexDataweb/default.html?Instrument={}".format(
-                self._get_instrument_name()))
-
-    @task("Update Instrument List")
-    def update_instlist(self):
-        """
-        Prompt user to add instrument to the list of known IBEX instruments
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Add the host name of the instrument to the list saved in the CS:INSTLIST PV")
-
-    @task("Update web dashboard")
-    def update_web_dashboard(self):
-        """
-        Prompt user to add the instrument to the web dashboard
-        """
-        redirect_page = os.path.join("C:", "inetpub", "wwwroot", "DataWeb", "Dashboards", "redirect.html")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Add the host name of the instrument to NDX_INSTS or ALL_INSTS in webserver.py in the JSON_bourne "
-            "repository.")
-        self.prompt.prompt_and_raise_if_not_yes(
-            "On NDAEXTWEB1, pull the updated code and add a link to the instrument dashboard on the main "
-            "dataweb page under {}".format(redirect_page))
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Restart JSON_bourne on NDAEXTWEB1 when appropriate. "
-            "(WARNING: This will kill all existing sessions!)")
-
-    @task("Update kafka topics")
-    def update_kafka_topics(self):
-        """
-        Adds the required kafka topics to the cluster.
-        """
-        add_required_topics("livedata.isis.cclrc.ac.uk:9092", self._get_instrument_name())
-
-    @task("Install wiring tables")
-    def install_wiring_tables(self):
-        """
-        Prompt user to install wiring tables in the appropriate folder.
-        """
-        tables_dir = os.path.join(SETTINGS_CONFIG_PATH, self._machine_name, "configurations", "tables")
-        self.prompt.prompt_and_raise_if_not_yes("Install the wiring tables in {}.".format(tables_dir))
-
-    @task("Configure motion setpoints")
-    def configure_motion(self):
-        """
-        Prompt user to configure Galils
-        """
-        url = "https://github.com/ISISComputingGroup/ibex_developers_manual/wiki/" \
-              "Deployment-on-an-Instrument-Control-PC#set-up-motion-set-points"
-        if self.prompt.prompt("Please configure the motion set points for this instrument. Instructions can be"
-                              " found on the developer wiki. Open instructions in browser now?",
-                              ["Y", "N"], "N") == "Y":
-            subprocess.call("explorer {}".format(url))
-        self.prompt.prompt_and_raise_if_not_yes("Confirm motion set points have been configured.")
-
-    @task("Add Nagios checks")
-    def add_nagios_checks(self):
-        """
-        Prompt user to add nagios checks.
-        """
-        # For future reference, genie_python can send emails!
-        self.prompt.prompt_and_raise_if_not_yes("Add this instrument to the Nagios monitoring system. Talk to "
-                                                "Freddie Akeroyd for help with this.")
-
-    @task("Inform instrument scientists")
-    def inform_instrument_scientists(self):
-        """
-        Inform instrument scientists that the machine has been upgraded.
-        """
-        # For future reference, genie_python can send emails!
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Inform the instrument scientists that the upgrade has been completed")
-
-    @task("Apply changes in release notes")
-    def apply_changes_noted_in_release_notes(self):
-        """
-        Apply any changes noted in the release notes.
-        """
-        # For future reference, genie_python can send emails!
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Look in the IBEX wiki at the release notes for the version you are deploying. Apply needed fixes.")
-
-    @contextmanager
-    def timestamped_pv_backups_file(self, name, directory, extension="txt"):
-        """
-        Context manager to create a timestamped file in the pv backups directory
-
-        Args:
-            name (str): path to the file
-            directory (str): directory of file
-            extension (str): the extension of the file
-        """
-        filename = os.path.join(PV_BACKUPS_DIR, directory, "{}_{}.{}"
-                                .format(name, datetime.today().strftime('%Y-%m-%d-%H-%M-%S'), extension))
-
-        if not os.path.exists(os.path.join(PV_BACKUPS_DIR, directory)):
-            os.makedirs(os.path.join(PV_BACKUPS_DIR, directory))
-
-        with open(filename, "w") as f:
-            yield f
-
-    @task("Save motor parameters to csv file")
-    def save_motor_parameters_to_file(self):
-        """
-        Saves the motor parameters to csv file.
-        """
-        with self.timestamped_pv_backups_file(name="motors", directory="motors", extension="csv") as f:
-            get_params_and_save_to_file(f)
-
-    @task("Save block parameters to file")
-    def save_blocks_to_file(self):
-        """
-        Saves block parameters in a file.
-        """
-
-        blocks = self._ca.get_blocks()
-
-        if blocks is None:
-            print("Blockserver unavailable - not archiving.")
-        else:
-            if blocks:
-                for block in blocks:
-                    with self.timestamped_pv_backups_file(name=block, directory="blocks") as f:
-                        f.write("{}\r\n".format(self._ca.cget(block)))
-            else:
-                print("Blockserver available but no blocks found - not archiving anything")
-
-    @task("Save blockserver PV to file")
-    def save_blockserver_pv_to_file(self):
-        """
-        Saves the blockserver PV to a file.
-        """
-
-        def _pretty_print(data):
-            return pprint.pformat(data, width=800, indent=2)
-
-        pvs_to_save = [
-            ("all_component_details", "CS:BLOCKSERVER:ALL_COMPONENT_DETAILS"),
-            ("block_names", "CS:BLOCKSERVER:BLOCKNAMES"),
-            ("groups", "CS:BLOCKSERVER:GROUPS"),
-            ("configs", "CS:BLOCKSERVER:CONFIGS"),
-            ("components", "CS:BLOCKSERVER:COMPS"),
-            ("runcontrol_out", "CS:BLOCKSERVER:GET_RC_OUT"),
-            ("runcontrol_pars", "CS:BLOCKSERVER:GET_RC_PARS"),
-            ("curr_config_details", "CS:BLOCKSERVER:GET_CURR_CONFIG_DETAILS"),
-            ("server_status", "CS:BLOCKSERVER:SERVER_STATUS"),
-            ("synoptic_names", "CS:SYNOPTICS:NAMES"),
-        ]
-
-        for name, pv in pvs_to_save:
-            with self.timestamped_pv_backups_file(directory="inst_servers", name=name) as f:
-                try:
-                    f.write("{}\r\n".format(_pretty_print(self._ca.get_object_from_compressed_hexed_json(pv))))
-                except Exception as e:
-                    print("Couldn't get data from {} because: {}".format(pv, e.message))
-                    f.write(e.message)
-
-    def check_resources(self):
-        """
-        Check the machine's resources meet minimum requirements.
-        """
-        self.check_virtual_memory()
-        self._check_disk_usage()
-
-    @task("Check virtual machine memory")
-    def check_virtual_memory(self):
-        """
-        Checks the machine's virtual memory meet minimum requirements.
-        """
-        ram = psutil.virtual_memory().total
-
-        machine_type = self.prompt.prompt(
-            "Is this machine an instrument (e.g. NDXALF) or a test machine (e.g. NDXSELAB)? (instrument/test) ",
-            possibles=["instrument", "test"], default="test")
-
-        if machine_type == "instrument":
-            min_memory = RAM_NORMAL_INSTRUMENT
-        else:
-            min_memory = RAM_MIN
-
-        if ram >= min_memory:
-            print("Virtual memory ({:.1f}GB) is already at or above the recommended level for this machine type "
-                  "({:.1f}GB). Nothing to do in this step.".format(ram/GIGABYTE, min_memory/GIGABYTE))
-        else:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "Current machine memory is {:.1f}GB, the recommended amount for this machine is {:.1f}GB.\n\n"
-                "If appropriate, upgrade this machine's memory allowance by following the instructions in "
-                "https://github.com/ISISComputingGroup/ibex_developers_manual/wiki/Increase-VM-memory.\n\n"
-                "Note, this will require a machine restart.".format(ram/GIGABYTE, min_memory/GIGABYTE)
-            )
-
-    @task("Check there is {:.1e}B free disk space".format(FREE_DISK_MIN))
-    def _check_disk_usage(self):
-        """
-        Checks the machine's free disk space meets minimum requirements.
-        """
-        disk_space = psutil.disk_usage("/")
-
-        if disk_space.free < FREE_DISK_MIN:
-            self.prompt.prompt_and_raise_if_not_yes(
-                "The machine requires at least {:.1f}GB of free disk space to run IBEX."
-                .format(FREE_DISK_MIN/GIGABYTE))
-
-    @task("Put IBEX autostart into pc start menu")
-    def put_autostart_script_in_startup_area(self):
-        """
-        Copies the ibex server autostart script into the PC startup folder so that the IBEX server starts
-        automatically on startup.
-        """
-
-        autostart_script_name = "ibex_system_boot.bat"
-
-        from_path = os.path.join(EPICS_PATH, autostart_script_name)
-        to_path = os.path.join(PC_START_MENU, "Programs", "Startup", autostart_script_name)
-
-        # Remove old version if exists
-        if os.path.exists(to_path):
-            try:
-                os.remove(to_path)
-            except (OSError, IOError):
-                self.prompt.prompt_and_raise_if_not_yes("Please manually remove file at '{}'".format(to_path))
-
-        try:
-            shutil.copyfile(from_path, to_path)
-        except (OSError, IOError):
-            self.prompt.prompt_and_raise_if_not_yes("Please manually copy file from '{}' to '{}'"
-                                                    .format(from_path, to_path))
-
-    @task("Restrict Internet Explorer")
-    def restrict_ie(self):
-        """
-        Restrict access of external websites to address a security vulnerability in Internet Explorer.
-        """
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Configure Internet Explorer to restrict access to the web except for select whitelisted sites:\n"
-            "- Open 'Internet Options' (from the gear symbol in the top right corner of the window).\n"
-            "- Go to the 'Connections' tab and open 'Lan Settings'\n"
-            "- Check 'Use Automatic configuration script' and enter http://dataweb.isis.rl.ac.uk/proxy.pac for 'Address'\n"
-            "- Click 'Ok' on all dialogs.")
-
-    @task("Change genie_python shortcuts to python 3")
-    def change_shortcuts_to_python_3(self):
-        """
-        Prompt user to find shortcuts to genie_python and replace them with Python 3 shortcuts
-        """
-
-        self.prompt.prompt_and_raise_if_not_yes(
-            "Relace any shortcuts to genie_python with those to python 3 in C:\\Instrument\\Apps\\Python3")
-
-    @task("Update Git")
-    def install_or_upgrade_git(self):
-        """
-        Install the latest git version
-        """
-        git_path = os.path.join(os.path.expanduser("~"), "AppData", "Local", "Programs", "Git", "cmd", "git.exe")
-        if os.path.exists(git_path):
-            os.system("{} update-git-for-windows".format(git_path))
-            self.prompt.prompt_and_raise_if_not_yes("Press Y/N if Git has installed correctly")
-        else:
-            self.prompt.prompt_and_raise_if_not_yes("Download and Install Git from https://git-scm.com/downloads")
-
-    @task("Initialize var dir")
-    def initialize_var_dir(self):
-        """
-        Creates the folder structure for the C:\instrument\var directory.
-        """
-        # config_env creates all the necessary directories for us
-        RunProcess(working_dir=EPICS_PATH, executable_file="config_env.bat").run()
-
-    @task("Copy VHDs to local area")
-    def copy_vhds_to_local_area(self):
-        if os.path.exists(LOCAL_VHD_DIR):
-            try:
-                shutil.rmtree(LOCAL_VHD_DIR)
-            except (IOError, OSError):
-                self.request_dismount_vhds()
-                shutil.rmtree(LOCAL_VHD_DIR)
-
-        os.mkdir(LOCAL_VHD_DIR)
-        for vhd in VHDS:
-            shutil.copyfile(os.path.join(REMOTE_VHD_SRC_DIR, vhd.source_filename),
-                            os.path.join(LOCAL_VHD_DIR, vhd.source_filename))
-
-    def _create_file_and_wait_for_it_to_be_deleted(self, filename, timeout):
-        with open(filename, "w") as f:
-            f.write("")
-
-        print("Waiting for file at {} to be deleted...".format(filename))
-        for _ in range(timeout):
-            if not os.path.exists(filename):
-                break
-            sleep(1)
-        else:
-            os.remove(filename)
-            raise IOError("File at {} still existed after {}s, check VHD scheduled task is running correctly"
-                          .format(filename, timeout))
-
-    @task("Request VHDs to be mounted")
-    def request_mount_vhds(self):
-        self._create_file_and_wait_for_it_to_be_deleted(FILE_TO_REQUEST_VHD_MOUNTING, VHD_MOUNT_DISMOUNT_TIMEOUT)
-
-    @task("Request VHDs to be dismounted")
-    def request_dismount_vhds(self):
-        self._create_file_and_wait_for_it_to_be_deleted(FILE_TO_REQUEST_VHD_DISMOUNTING, VHD_MOUNT_DISMOUNT_TIMEOUT)
-            
-    @task("Mount VHDs")
-    def mount_vhds(self):
-        if not os.path.exists(FILE_TO_REQUEST_VHD_MOUNTING):
-            return
-
-        admin_commands = AdminCommandBuilder()
-
-        admin_commands.add_command("sc", "stop MYSQL80", expected_return_val=None)
-
-        for vhd in VHDS:
-            driveletter_file = os.path.join(tempfile.gettempdir(), "{}_driveletter.txt".format(vhd.name))
-            if os.path.exists(driveletter_file):
-                os.remove(driveletter_file)
-
-            if os.path.exists(vhd.mount_point):
-                admin_commands.add_command("move",
-                                           '"{mount_point}" "{mount_point}_backup"'.format(mount_point=vhd.mount_point),
-                                           expected_return_val=None)
-
-            # Mount the VHD and write it's assigned drive letter to a file.
-            admin_commands.add_command(
-                "powershell",
-                r'-command "Hyper-V\Mount-VHD -path {vhd_file} -Passthru | Get-Disk | Get-Partition | Get-Volume | foreach {{ $_.DriveLetter }} | out-file -filepath {driveletter_file} -Encoding ASCII -NoNewline"'
-                    .format(vhd_file=os.path.join(LOCAL_VHD_DIR, vhd.source_filename), name=vhd.name, driveletter_file=driveletter_file))
-
-            # Append :\\ to drive letter, e.g. E -> E:\\ (this is necessary so that directory junctions work correctly)
-            admin_commands.add_command(
-                "powershell",
-                r'-command "echo :\\ | out-file -filepath {driveletter_file} -Encoding ASCII -Append -NoNewline"'
-                    .format(driveletter_file=driveletter_file)
-            )
-
-            # Create a directory junction from the mount point to the disk's assigned drive letter
-            admin_commands.add_command(
-                "powershell",
-                r'-command "&cmd /c mklink /J /D {mount_point} @(cat {driveletter_file})"'
-                    .format(mount_point=vhd.mount_point, name=vhd.name, driveletter_file=driveletter_file))
-
-        admin_commands.run_all()
-
-        os.remove(FILE_TO_REQUEST_VHD_MOUNTING)
-
-    @task("Dismount VHDs")
-    def dismount_vhds(self):
-
-        if not os.path.exists(FILE_TO_REQUEST_VHD_DISMOUNTING):
-            return
-
-        admin_commands = AdminCommandBuilder()
-
-        for vhd in VHDS:
-            # Dismount the VHD
-            admin_commands.add_command(
-                "powershell",
-                r'-command "Hyper-V\Dismount-VHD {vhd_file}"'.format(vhd_file=os.path.join(LOCAL_VHD_DIR, vhd.source_filename)),
-                expected_return_val=None,
-            )
-
-            # Remove directory junction
-            admin_commands.add_command(
-                "cmd",
-                r'/c "rmdir {mount_point}"'.format(mount_point=vhd.mount_point),
-                expected_return_val=None,
-            )
-
-            # Restore previous directories to where they were before mounting VHDS
-            if os.path.exists("{}_backup".format(vhd.mount_point)):
-                admin_commands.add_command("move",
-                                           '"{mount_point}_backup" "{mount_point}"'.format(mount_point=vhd.mount_point),
-                                           expected_return_val=None)
-
-        admin_commands.add_command("sc", "start MYSQL80", expected_return_val=None)
-        admin_commands.run_all()
-
-        os.remove(FILE_TO_REQUEST_VHD_DISMOUNTING)
-
-    @task("Deploy VHDS")
-    def deploy_vhds(self):
-
-        build_folder = os.path.join(REMOTE_VHD_DEST_DIR, "Build{}".format(os.environ["BUILD_NUMBER"]))
-        os.makedirs(build_folder)
-
-        for vhd in VHDS:
-            print("Deploying {} to '{}'".format(vhd.source_filename, build_folder))
-            shutil.copyfile(os.path.join(LOCAL_VHD_DIR, vhd.source_filename), os.path.join(build_folder, vhd.dest_filename))
-
-        print("Cleaning up local artefacts...")
-        shutil.rmtree(LOCAL_VHD_DIR)
