+++ installation_and_upgrade/ibex_install_utils/tasks/backup_tasks.py
@@ -1,0 +1,74 @@
+import os
+import shutil
+
+
+from ibex_install_utils.task import task
+from ibex_install_utils.tasks import BaseTasks
+from ibex_install_utils.tasks.common_paths import INSTRUMENT_BASE_DIR, BACKUP_DATA_DIR, BACKUP_DIR, EPICS_PATH, \
+    PYTHON_PATH, PYTHON_3_PATH, EPICS_UTILS_PATH, GUI_PATH, GUI_PATH_E4
+
+ALL_INSTALL_DIRECTORIES = (EPICS_PATH, PYTHON_PATH, PYTHON_3_PATH, GUI_PATH, GUI_PATH_E4, EPICS_UTILS_PATH)
+
+
+class BackupTasks(BaseTasks):
+
+    def _backup_dir(self, src, copy=True):
+        backup_dir = os.path.join(self._get_backup_dir(), os.path.basename(src))
+        if src in os.getcwd():
+            self.prompt.prompt_and_raise_if_not_yes(
+                "You appear to be trying to delete the folder, {}, containing the current working directory {}. "
+                "Please do this manually to be on the safe side".format(src, os.getcwd()))
+        elif os.path.exists(backup_dir):
+            self.prompt.prompt_and_raise_if_not_yes(
+                "Backup dir {} already exist. Please backup this app manually".format(backup_dir))
+        elif os.path.exists(src):
+            if copy:
+                print("Copying {} to {}".format(src, backup_dir))
+                shutil.copytree(src, backup_dir)
+            else:
+                print("Moving {} to {}".format(src, backup_dir))
+                self._file_utils.move_dir(src, backup_dir, self.prompt)
+
+    @task("Backup old directories")
+    def backup_old_directories(self):
+        """
+        Backup old directories
+        """
+        if os.path.exists(BACKUP_DATA_DIR):
+            if not os.path.exists(BACKUP_DIR):
+                os.mkdir(BACKUP_DIR)
+
+            # Delete all but the oldest backup
+            current_backups = [os.path.join(BACKUP_DIR, d) for d in os.listdir(BACKUP_DIR)
+                               if os.path.isdir(os.path.join(BACKUP_DIR, d)) and d.startswith("ibex_backup")]
+            if len(current_backups) > 0:
+                all_but_newest_backup = sorted(current_backups, key=os.path.getmtime)[:-1]
+                backups_to_delete = all_but_newest_backup
+            else:
+                backups_to_delete = []
+
+            for d in backups_to_delete:
+                print("Removing backup {}".format(d))
+                self._file_utils.remove_tree(os.path.join(BACKUP_DIR, d), self.prompt)
+
+            # Move the folders
+            for app_path in ALL_INSTALL_DIRECTORIES:
+                self._backup_dir(app_path, copy=False)
+
+            # Backup settings and autosave
+            self._backup_dir(os.path.join(INSTRUMENT_BASE_DIR, "Settings"))
+            self._backup_dir(os.path.join(INSTRUMENT_BASE_DIR, "var", "Autosave"))
+        else:
+            self.prompt.prompt_and_raise_if_not_yes(
+                "Unable to find data directory '{}'. Please backup the current installation of IBEX "
+                "manually".format(BACKUP_DATA_DIR))
+
+    @task("Removing old version of IBEX")
+    def remove_old_ibex(self):
+        """
+        Removes older versions of IBEX server, client, genie_python and epics utils
+        Returns:
+
+        """
+        for path in ALL_INSTALL_DIRECTORIES:
+            self._file_utils.remove_tree(path, self.prompt)
